[Gold V] Title: 다각형의 면적, Time: 0 ms, Memory: 2180 KB -BaekjoonHub

+++ 백준/Gold/2166. 다각형의 면적/다각형의 면적.cc
@@ -6,7 +6,7 @@ using namespace std;
 using ll = long long;
 using P = pair <ll, ll>;
 
-double CalcExtent(const P& a, const P& b)
+ll CalcExtent(const P& a, const P& b)
 {
 	return a.first * b.second - a.second * b.first;
 }
